Re-implemented the observation alignment algorithm.
1. The previous implementation ended up being slow.
2. Now
   a) Sort all the epochs where any observations are available
   b) Compute the number of epochs between each adjacent pair of
   observations.
   c) Find the most frequent gap between adjacent observations. If more
   than on gap has the same highest frequency, get the smallest of
   those.
   d) Choose that are the modeling period (or observation sampling
   period, number of epochs per single modeling delta t).
   e) Convert observation epochs to modeling timestamps based on the
   selected modeling period.

+++ lib/prediction.cpp
@@ -18,7 +18,7 @@ using fmt::print;
 */
 
 void AnalysisGraph::generate_latent_state_sequences(
-    int initial_prediction_step) {
+    double initial_prediction_step) {
 
   // Allocate memory for prediction_latent_state_sequences
   this->predicted_latent_state_sequences.clear();
